requirements.txt
PyQt5==5.15.9
Pillow==9.5.0
PyAutoGUI==0.9.54
python-dotenv==1.0.0
mss==9.0.1 